requirements.txt
biopython==1.78
blas==1.0
ca-certificates==2020.12.5
certifi==2020.12.5
intel-openmp==2019.4
libcxx==10.0.0
libedit==3.1.20191231
libffi==3.3
mkl==2019.4=233
mkl-service==2.3.0
mkl_fft==1.2.0
mkl_random==1.1.1
ncurses==6.2
numpy==1.19.2
numpy-base==1.19.2
openssl==1.1.1i
pandas==1.1.5
pip==20.3.3
python==3.7.9
python-dateutil==2.8.1
python_abi==3.7
pytz==2020.4
readline==8.0
setuptools==51.0.0
six==1.15.0
sqlite==3.33.0
tk==8.6.10
wheel==0.36.2
xz==5.2.5
zlib==1.2.11
